+++ execution_engine/src/core/runtime/mod.rs
@@ -1493,6 +1493,8 @@ where
             phase,
             self.config,
             transfers,
+            // Pass on this deploy's spending limit to the inner context.
+            *self.context().main_purse_spending_limit(),
         );
 
         let mut mint_runtime = self.new_from_self(mint_context, stack);
@@ -1581,6 +1583,10 @@ where
         // Result still contains a result, but the entrypoints logic does not exit early on errors.
         let ret = result?;
 
+        // Update outer spending approved limit.
+        self.context
+            .set_spending_limit(*mint_runtime.context.main_purse_spending_limit());
+
         let urefs = extract_urefs(&ret)?;
         let access_rights = extract_access_rights_from_urefs(urefs);
         self.context.access_rights_extend(access_rights);
@@ -1639,6 +1645,7 @@ where
             phase,
             self.config,
             transfers,
+            *self.context.main_purse_spending_limit(),
         );
 
         let mut runtime = self.new_from_self(runtime_context, stack);
@@ -1759,6 +1766,8 @@ where
             phase,
             self.config,
             transfers,
+            // We should not spend any tokens when executing auction contract.
+            U512::zero(),
         );
 
         let mut runtime = self.new_from_self(runtime_context, stack);
@@ -2244,6 +2253,7 @@ where
             self.context.phase(),
             self.config,
             self.context.transfers().to_owned(),
+            *self.context().main_purse_spending_limit(),
         );
 
         let mut stack = self.stack.clone();
@@ -2289,6 +2299,8 @@ where
                     // running session code
                     *self.context.named_keys_mut() = runtime.context.named_keys().clone();
                 }
+                self.context
+                    .set_spending_limit(*runtime.context.main_purse_spending_limit());
                 return Ok(runtime.take_host_buffer().unwrap_or(CLValue::from_t(())?));
             }
         };