+++ execution_engine_testing/tests/src/test/contract_api/get_call_stack.rs
@@ -2507,7 +2507,7 @@ mod payment {
         DeployItemBuilder, ExecuteRequestBuilder, DEFAULT_ACCOUNT_ADDR,
     };
     use casper_execution_engine::shared::wasm;
-    use casper_types::{runtime_args, RuntimeArgs};
+    use casper_types::{runtime_args, system::mint, RuntimeArgs, U512};
 
     use super::{
         AccountExt, ARG_CALLS, ARG_CURRENT_DEPTH, CONTRACT_CALL_RECURSIVE_SUBCALL,
@@ -2517,6 +2517,16 @@ mod payment {
     // DEPTHS should not contain 1, as it will eliminate the initial element from the subcalls
     // vector.  Going further than 5 will git the gas limit.
     const DEPTHS: &[usize] = &[0, 2, 5];
+
+    // Constant from the contracts used in the tests below.
+    const LARGE_AMOUNT: u64 = 1_500_000_000_000;
+
+    // In the payment phase, this test will try to transfer `len + 1` times
+    // a fixed amount of `1_500_000_000_000` from the main purse of the account.
+    // We need to provide an explicit approval via passing that as an `amount` argument.
+    fn approved_amount(idx: usize) -> U512 {
+        U512::from(LARGE_AMOUNT * (idx + 1) as u64)
+    }
 
     // Session + recursive subcall
 
@@ -2546,6 +2556,7 @@ mod payment {
                 let args = runtime_args! {
                     ARG_CALLS => subcalls.clone(),
                     ARG_CURRENT_DEPTH => 0u8,
+                    mint::ARG_AMOUNT => approved_amount(*len),
                 };
                 let deploy = DeployItemBuilder::new()
                     .with_address(sender)
@@ -2592,6 +2603,7 @@ mod payment {
                 let args = runtime_args! {
                     ARG_CALLS => subcalls.clone(),
                     ARG_CURRENT_DEPTH => 0u8,
+                    mint::ARG_AMOUNT => approved_amount(*len),
                 };
                 let deploy = DeployItemBuilder::new()
                     .with_address(sender)
@@ -2637,6 +2649,7 @@ mod payment {
                 let args = runtime_args! {
                     ARG_CALLS => subcalls.clone(),
                     ARG_CURRENT_DEPTH => 0u8,
+                    mint::ARG_AMOUNT => approved_amount(*len),
                 };
                 let deploy = DeployItemBuilder::new()
                     .with_address(sender)
@@ -2680,6 +2693,7 @@ mod payment {
                 let args = runtime_args! {
                     ARG_CALLS => subcalls.clone(),
                     ARG_CURRENT_DEPTH => 0u8,
+                    mint::ARG_AMOUNT => approved_amount(*len),
                 };
                 let deploy = DeployItemBuilder::new()
                     .with_address(sender)
@@ -2729,6 +2743,7 @@ mod payment {
                 let args = runtime_args! {
                     ARG_CALLS => subcalls.clone(),
                     ARG_CURRENT_DEPTH => 0u8,
+                    mint::ARG_AMOUNT => approved_amount(*len),
                 };
                 let deploy = DeployItemBuilder::new()
                     .with_address(sender)
@@ -2771,6 +2786,7 @@ mod payment {
                 let args = runtime_args! {
                     ARG_CALLS => subcalls.clone(),
                     ARG_CURRENT_DEPTH => 0u8,
+                    mint::ARG_AMOUNT => approved_amount(*len),
                 };
                 let deploy = DeployItemBuilder::new()
                     .with_address(sender)
@@ -2812,6 +2828,7 @@ mod payment {
                 let args = runtime_args! {
                     ARG_CALLS => subcalls.clone(),
                     ARG_CURRENT_DEPTH => 0u8,
+                    mint::ARG_AMOUNT => approved_amount(*len),
                 };
                 let deploy = DeployItemBuilder::new()
                     .with_address(sender)
@@ -2850,6 +2867,7 @@ mod payment {
                 let args = runtime_args! {
                     ARG_CALLS => subcalls.clone(),
                     ARG_CURRENT_DEPTH => 0u8,
+                    mint::ARG_AMOUNT => approved_amount(*len),
                 };
                 let deploy = DeployItemBuilder::new()
                     .with_address(sender)